Arreglando errors con mjs y reorganizando carpetas

+++ AICHAT-CanchasFutbol/src/main/java/JMR/AICHAT/Controller/MensajeController.java
@@ -1,9 +1,8 @@
 package JMR.AICHAT.Controller;
 
 
-import JMR.AICHAT.DTOs.DatosMensaje;
-import JMR.AICHAT.DTOs.Outputs.MensajeResponse;
-import JMR.AICHAT.Mapper.MensajeMapper;
+import JMR.AICHAT.Mensaje.MensajeResponse;
+import JMR.AICHAT.Mensaje.MensajeMapper;
 import JMR.AICHAT.Mensaje.Mensaje;
 import JMR.AICHAT.Mensaje.MensajeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +10,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -40,7 +38,7 @@ public class MensajeController {
         return ResponseEntity.ok(respuesta);
     }
 
-    @GetMapping("/intencion")
+    @GetMapping("/intencion/{intencion}")
     public ResponseEntity<List<MensajeResponse>>  ObtenerTodosPorIntencion(@PathVariable String intencion){
         List<Mensaje> mensajes = mensajeRepository.findByIntencion(intencion);
         List<MensajeResponse> respuesta= mensajes.stream()
@@ -50,24 +48,27 @@ public class MensajeController {
     }
 
     @Transactional
-    @DeleteMapping("/{id}")
+    @DeleteMapping("/id/{id}")
     public ResponseEntity eliminarMensaje(@PathVariable Long id){
+        if (!mensajeRepository.existsById(id)) {
+            throw new RuntimeException("Mensaje no encontrado");
+        }
         mensajeRepository.deleteById(id);
-        return ResponseEntity.ok().build();
+        return ResponseEntity.noContent().build();
     }
 
     @Transactional
-    @DeleteMapping("/{telefono}")
+    @DeleteMapping("/telefono/{telefono}")
     public ResponseEntity eliminarMensajePorTelefono(@PathVariable String telefono){
         mensajeRepository.deleteByTelefono(telefono);
-        return ResponseEntity.ok().build();
+        return ResponseEntity.noContent().build();
     }
 
     @Transactional
     @DeleteMapping("/todos")
     public ResponseEntity eliminarTodos(){
         mensajeRepository.deleteAll();
-        return ResponseEntity.ok().build();
+        return ResponseEntity.noContent().build();
     }
